requirements.txt
Flask==3.0.0
Flask-SQLAlchemy==3.1.1
PyMySQL==1.1.0
bcrypt==4.1.2
python-dotenv==1.0.0
cryptography==42.0.5
numpy==2.2.6
tensorflow==2.20.0
Pillow==12.1.1
flask-cors==6.0.2
pandas==2.3.3
SQLAlchemy==2.0.47
h5py==3.15.1
